+++ packages/web/src/framework/web.ts
@@ -23,7 +23,7 @@ import { loggers, MidwayContextLogger } from '@midwayjs/logger';
 import { resolve } from 'path';
 import { Server } from 'net';
 import { debuglog } from 'util';
-import { getFreePort } from '../utils';
+import { AddressInfo, createServer } from 'net';
 
 const debug = debuglog('midway:debug');
 
@@ -338,4 +338,19 @@ export class MidwayWebFramework extends BaseFramework<
   public setServer(server) {
     this.server = server;
   }
+}
+
+async function getFreePort() {
+  return new Promise<number>((resolve, reject) => {
+    const server = createServer();
+    server.listen(0, () => {
+      try {
+        const port = (server.address() as AddressInfo).port;
+        server.close();
+        resolve(port);
+      } catch (err) {
+        reject(err);
+      }
+    });
+  });
 }
